fixed error:  would bomb out if .gemrc existed

+++ recipes/gem_no_rdoc_no_ri.rb
@@ -6,7 +6,9 @@ file "/etc/gemrc" do
   content "install: --no-rdoc --no-ri\nupdate: --no-rdoc --no-ri\n"
 end
 
-execute "link .gemrc to /etc/gemrc" do
-  command "ln -s /etc/gemrc #{WS_HOME}/.gemrc"
-  user WS_USER
+unless File.exists?("#{WS_HOME}/.gemrc")
+  execute "link .gemrc to /etc/gemrc" do
+    command "ln -s /etc/gemrc #{WS_HOME}/.gemrc"
+    user WS_USER
+  end
 end
